requirements.txt
python-telegram-bot==20.7
sqlite3
python-dotenv==1.0.0
cryptography==41.0.7
qrcode==7.4.2
pillow==10.1.0
requests==2.31.0